requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.5.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
chromadb==0.3.14
click==8.1.3
clickhouse-connect==0.5.17
commonmark==0.9.1
dataclasses-json==0.5.7
Deprecated==1.2.13
duckdb==0.7.1
et-xmlfile==1.1.0
fastapi==0.95.0
ffmpeg-python==0.2.0
filelock==3.10.7
frozenlist==1.3.3
future==0.18.3
geojson==2.5.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.147
llvmlite==0.39.1
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
openai==0.27.2
openai-whisper @ git+https://github.com/openai/whisper.git@b5851c6c40e753606765ac45b85b298e3ae9e00d
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
posthog==2.4.0
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.14.0
pyowm==3.3.0
pypandoc==1.11
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytube==12.1.3
pytz==2023.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.1
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.3
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.7
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
youtube-search-python==1.6.6
yt-dlp==2023.3.4
zipp==3.15.0
zstandard==0.20.0
